refactor(core): make core really work with babel7
