--- packages/core/lib/env.js
@@ -1,27 +1,0 @@
-'use strict';
-
-const isPlainObject = require('is-plain-object');
-
-exports.environmentalize = (_config) => {
-  const _env = {};
-  const env = global._env || process.env;
-  const regExp = /\[(\w+?)\]/g;
-  const replaceRecursive = (item) => {
-    if (Array.isArray(item)) {
-      return item.map(replaceRecursive);
-    }
-    if (isPlainObject(item)) {
-      return Object.keys(item).reduce((result, key) => {
-        result[key] = replaceRecursive(item[key]);
-        return result;
-      }, {});
-    }
-    if (regExp.test(item)) {
-      return item.replace(regExp, (_, key) =>
-        replaceRecursive((_env[key] = env[key] || ''))
-      );
-    }
-    return item;
-  };
-  return Object.assign(replaceRecursive(_config), { _config, _env });
-};
