+++ packages/core/lib/runtime.js
@@ -1,19 +1,45 @@
 'use strict';
 
-const debug = require('debug')('untool:runtime');
-const define = require('mixinable');
+import createDebug from 'debug';
+import define from 'mixinable';
 
-const { environmentalize } = require('./env');
+import isPlainObject from 'is-plain-object';
+
+const debug = createDebug('untool:runtime');
 
 const { override } = define;
 
-exports.Mixin = class Mixin {
+const environmentalize = (_config) => {
+  const _env = {};
+  const env = global._env || process.env;
+  const regExp = /\[(\w+?)\]/g;
+  const replaceRecursive = (item) => {
+    if (Array.isArray(item)) {
+      return item.map(replaceRecursive);
+    }
+    if (isPlainObject(item)) {
+      return Object.keys(item).reduce((result, key) => {
+        result[key] = replaceRecursive(item[key]);
+        return result;
+      }, {});
+    }
+    if (regExp.test(item)) {
+      return item.replace(regExp, (_, key) =>
+        replaceRecursive((_env[key] = env[key] || ''))
+      );
+    }
+    return item;
+  };
+  return Object.assign(replaceRecursive(_config), { _config, _env });
+};
+
+export class Mixin {
   constructor(config) {
     this.config = config;
   }
-};
+}
 
-exports.render = function render(...renderArgs) {
+export function render(...renderArgs) {
   return (baseConfig, mixins) => {
     const config = environmentalize(baseConfig);
     const strategies = mixins.reduce(
@@ -28,4 +54,4 @@ exports.render = function render(...renderArgs) {
       createMixinable(config, ...renderArgs).render(...callArgs);
     };
   };
-};
+}
